+++ src/query/lib/index.ts
@@ -3,3 +3,4 @@ export * from "./broadcastSyncDriver";
 export * from "./stableStringify";
 export * from "./toKeyed";
 export * from "./withAbort";
+export * from "./wrapTrigger";