+++ src/query/react/useCommand.ts
@@ -1,13 +1,13 @@
 import React from "react";
 
 import { useConstant, useEventHandler } from "@/common/react";
-import type { ICommand, TCommandAgentState } from "@/query/types";
+import type { ICommand, TCommandAgentState, TTriggerPromise } from "@/query/types";
 import { useSignal } from "@/signals/react";
 
 export function useCommand<TArgs, TData>(
     command: ICommand<TArgs, TData>,
     key?: string,
-): [trigger: (args: TArgs) => Promise<TData>, state: TCommandAgentState<TArgs, TData>] {
+): [trigger: (args: TArgs) => TTriggerPromise<TData>, state: TCommandAgentState<TArgs, TData>] {
     const agent = useConstant(() => command.createAgent(key), [command]);
 
     React.useEffect(() => {